% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed2020.R
\docType{data}
\name{ed2020}
\alias{ed2020}
\title{National Hospital Ambulatory Medical Care Survey (NHAMCS) Emergency Department (ED) 2020 Public Use File (PUF)}
\format{
An object of class \code{survey.design2} (inherits from \code{survey.design}) with 14860 rows and 1018 columns.
}
\source{
\itemize{
\item SAS data: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NHAMCS/sas/ed2020_sas.zip}
\item Survey design variables: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NHAMCS/sas/readme2020-ed-sas.txt}
\item SAS formats: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NHAMCS/sas/ed20for.txt}
\item Documentation: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NHAMCS/doc20-ed-508.pdf}
\item Summary Tables: \url{https://www.cdc.gov/nchs/data/nhamcs/web_tables/2020-nhamcs-ed-web-tables-508.pdf}
}
}
\usage{
ed2020
}
\description{
A data system of visits to emergency departments. Note that the unit of
observation is visits, not patients – this distinction is important since a single
patient can make multiple visits.
}
\keyword{datasets}
