% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchsdata-package.R
\docType{package}
\name{nchsdata-package}
\alias{nchsdata}
\alias{nchsdata-package}
\title{nchsdata: Public use files (PUFs) from the National Center for Health Statistics (NCHS)}
\description{
Public use files (PUFs) from the National Center for Health Statistics (NCHS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cdcgov.github.io/nchsdata}
  \item \url{https://github.com/CDCgov/nchsdata}
}

}
\author{
\strong{Maintainer}: Alex Strashny \email{AStrashny@cdc.gov}

}
\keyword{internal}
