% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes2016s.R
\docType{data}
\name{nhanes2016s}
\alias{nhanes2016s}
\title{National Health and Nutrition Examination Survey (NHANES) 2013-2016 Public Use File (PUF) (select variables and observations)}
\format{
An object of class \code{survey.design2} (inherits from \code{survey.design}) with 9942 rows and 13 columns.
}
\source{
\itemize{
\item SAS Transport Files / Demographic (DEMO) (2013-2014): \url{https://wwwn.cdc.gov/nchs/nhanes/2013-2014/DEMO_H.XPT}
\item SAS Transport Files / Demographic (DEMO) (2015-2016): \url{https://wwwn.cdc.gov/nchs/nhanes/2015-2016/DEMO_I.XPT}
\item SAS Transport Files / Mental Health - Depression Screener (DPQ) (2013-2014): \url{http://wwwn.cdc.gov/nchs/nhanes/2013-2014/DPQ_H.XPT}
\item SAS Transport Files / Mental Health - Depression Screener (DPQ) (2015-2016): \url{http://wwwn.cdc.gov/nchs/nhanes/2015-2016/DPQ_I.XPT}
\item Documentation: \url{https://wwwn.cdc.gov/nchs/nhanes/}
\item "Prevalence of Depression Among Adults Aged 20 and Over: United States, 2013-2016": \url{https://www.cdc.gov/nchs/products/databriefs/db303.htm}
\item Example code for the above report: \url{https://wwwn.cdc.gov/nchs/data/Tutorials/Code/DB303_R.r}
}
}
\usage{
nhanes2016s
}
\description{
A data system of the health and nutritional status of adults and children across the United States. The variables and observations selected are those used in the following report: Brody DJ, Pratt LA, Hughes JP. Prevalence of Depression Among Adults Aged 20 and Over: United States, 2013-2016. NCHS Data Brief. No 303. Hyattsville, MD: National Center for Health Statistics. 2018. \url{https://www.cdc.gov/nchs/products/databriefs/db303.htm}
}
\keyword{datasets}
