% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc2018.R
\docType{data}
\name{rcc2018}
\alias{rcc2018}
\title{National Study of Long-Term Care Providers (NSLTCP) Residential Care Community (RCC) 2018 Public Use File (PUF)}
\format{
An object of class \code{survey.design2} (inherits from \code{survey.design}) with 503 rows and 155 columns.
}
\source{
\itemize{
\item SAS data: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Datasets/NPALS/final2018rcc_puf.sas7bdat}
\item Documentation: \url{https://www.cdc.gov/nchs/data/npals/2018NSLTCP-RCC-PUF-ReadMeFile_provider.pdf}
\item Codebook: \url{https://www.cdc.gov/nchs/data/npals/final2018rcc_puf_codebook.pdf}
}
}
\usage{
rcc2018
}
\description{
A data system of RCCs.
}
\keyword{datasets}
